refactor: 공통 컴포넌트 삭제 및 공통 스타일링 별도의 파일로 분리
