+++ src/features/my-document/components/Table.jsx
@@ -1,11 +1,11 @@
 import React, { useEffect, useState } from 'react';
-import { collection, query, getDocs, where, orderBy } from 'firebase/firestore';
 import { db } from '@/shared/api/firebase/firebase'
+import { collection, query, getDocs, where, orderBy } from 'firebase/firestore';
 import { TableHeader } from './TableHeader';
 import { TableRow } from './TableRow.jsx';
-import { ExpandedRow } from './ExpandedRow';
-import { Pagination } from './Pagination';
-import { Container, TableWrap, TableContent, Tbody } from '../TableStyles';
+import { ExpandedRow } from '@/features/admin/approval/components/ExpandedRow';
+import { Pagination } from '@/shared/components/pagination/Pagination';
+import * as Common from '@/features/admin/approval/TableCommonStyles';
 
 export function Table({ filterValue }) {
   const [data, setData] = useState([]); // 전체 데이터 저장
@@ -94,12 +94,12 @@ export function Table({ filterValue }) {
 
   return (
     <>
-      <Container>
-        <TableWrap>
-          <TableContent>
+      <Common.Container>
+        <Common.TableWrap>
+          <Common.TableContent>
             {/* 테이블 헤더 */}
             <TableHeader />
-            <Tbody>
+            <Common.Tbody>
               {/* 현재 페이지의 데이터를 렌더링 */}
               {getCurrentPageData().map((item) => (
                 <React.Fragment key={item.id}>
@@ -113,10 +113,10 @@ export function Table({ filterValue }) {
                   {expandedId === item.id && <ExpandedRow content={item.requestContent} />}
                 </React.Fragment>
               ))}
-            </Tbody>
-          </TableContent>
-        </TableWrap>
-      </Container>
+            </Common.Tbody>
+          </Common.TableContent>
+        </Common.TableWrap>
+      </Common.Container>
       {/* 페이지네이션 */}
       <Pagination
         totalItems={filteredData.length} // 전체 항목 수 전달
